Refactored Handler Code to Reduce Duplication
- Moved shared functions to separate file
- Added App Presenter to match Package Presenter
