+++ cfshim/handlers/shared_functions.go
@@ -1,8 +1,14 @@
 package handlers
 
 import (
+	"context"
 	"encoding/json"
+	"fmt"
 	"net/http"
+
+	appsv1alpha1 "cloudfoundry.org/cf-crd-explorations/api/v1alpha1"
+	"cloudfoundry.org/cf-crd-explorations/cfshim/filters"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 func ReturnFormattedError(w http.ResponseWriter, status int, title string, detail string, code int) {
@@ -17,4 +23,29 @@ func ReturnFormattedError(w http.ResponseWriter, status int, title string, detai
 			},
 		},
 	})
+}
+
+// getAppListFromQuery takes URL query parameters and queries the K8s Client for all Apps
+// builds a filter based on params and walks through, placing every match into the returned list of Apps
+// returns an error if something went wrong with the K8s query
+func getAppListFromQuery(c *client.Client, queryParameters map[string][]string) ([]*appsv1alpha1.App, error) {
+	var filter Filter = &filters.AppFilter{
+		QueryParameters: queryParameters,
+	}
+
+	// Get all the CF Apps from K8s API store in AllApps which contains Items: []App
+	AllApps := &appsv1alpha1.AppList{}
+	err := (*c).List(context.Background(), AllApps)
+	if err != nil {
+		return nil, fmt.Errorf("error fetching app: %v", err)
+	}
+
+	// Apply filter to AllApps and store result in matchedApps
+	var matchedApps []*appsv1alpha1.App
+	for i, _ := range AllApps.Items {
+		if filter.Filter(&AllApps.Items[i]) {
+			matchedApps = append(matchedApps, &AllApps.Items[i])
+		}
+	}
+	return matchedApps, nil
 }
